refactor: Cria a pasta url para armazenar a BASE_URL.
O objetivo é tentar eliminar o conflito de não reconhecimento de path no deploy do front.
