+++ frontend/src/components/Form/index.js
@@ -4,7 +4,7 @@ import { useRequestData } from "../../hooks/useRequestData";
 import { useForm } from "../../hooks/useForm";
 import { useNavigate } from "react-router-dom";
 import { goToEndPage } from "../../routes/Coordinator";
-import { BASE_URL } from "../../constants/url.JS";
+import { BASE_URL } from "../../constants/url/baseUrl";
 
 function Form({ productList, setProductList }) {
   const navigate = useNavigate();
